scripts/debug/diagnose_training.py
#!/usr/bin/env python3
"""
Diagnose training issues and provide recommendations
"""

import os
import numpy as np
from dotenv import load_dotenv

load_dotenv()

# Analyze the training results
def analyze_training_metrics():
    print("=== Options Trading Bot Training Analysis ===\n")
    
    # Given metrics
    avg_reward = 13.50
    avg_loss = 5967.8686
    initial_capital = 100000
    episodes = 60
    
    # Calculate returns
    total_pnl = avg_reward * episodes
    pct_return = (total_pnl / initial_capital) * 100
    daily_return = pct_return / (episodes / 390)  # Assuming 390 steps per day
    
    print(f"1. RETURNS ANALYSIS:")
    print(f"   - Average P&L per episode: ${avg_reward:.2f}")
    print(f"   - Total P&L after {episodes} episodes: ${total_pnl:.2f}")
    print(f"   - Percentage return: {pct_return:.4f}%")
    print(f"   - Estimated daily return: {daily_return:.4f}%")
    print(f"   - Annualized return: {daily_return * 252:.2f}%")
    
    print(f"\n2. ISSUES IDENTIFIED:")
    print(f"   ❌ Returns are too low (${avg_reward} per episode = 0.0135% on $100k)")
    print(f"   ❌ High training loss ({avg_loss:.2f}) indicates learning difficulties")
    print(f"   ❌ Price data shows $100 instead of real SPY prices (~$600)")
    print(f"   ❌ Zero volatility in price history")
    
    print(f"\n3. ROOT CAUSES:")
    print(f"   - The agent is barely trading (too conservative)")
    print(f"   - Reward scaling is too small (needs 10-100x increase)")
    print(f"   - Price history data is not properly loaded")
    print(f"   - Position sizes might be too small")
    
    print(f"\n4. RECOMMENDATIONS:")
    print(f"   a) Fix data pipeline:")
    print(f"      - Ensure real stock prices are used in observations")
    print(f"      - Verify price history has actual volatility")
    print(f"   b) Adjust reward scaling:")
    print(f"      - Multiply portfolio change reward by 10-50x")
    print(f"      - Add bonus rewards for profitable trades")
    print(f"   c) Increase position sizes:")
    print(f"      - Allow larger positions (5-10% of capital)")
    print(f"      - Reduce transaction cost penalties")
    print(f"   d) Normalize inputs:")
    print(f"      - Normalize prices by dividing by 100")
    print(f"      - Normalize volumes and Greeks")
    
    print(f"\n5. EXPECTED GOOD RESULTS:")
    print(f"   ✓ Average reward: $500-2000 per episode (0.5-2% returns)")
    print(f"   ✓ Training loss: < 100 after convergence")
    print(f"   ✓ Daily returns: 0.5-2% (realistic for options)")
    print(f"   ✓ Win rate: 40-60%")
    
    print(f"\n6. QUICK FIXES TO TRY:")
    print(f"   - In options_trading_env.py: Change 'portfolio_change * 2.0' to 'portfolio_change * 20.0'")
    print(f"   - In historical_options_data.py: Fix price history to use real stock prices")
    print(f"   - Increase position sizes in _execute_action")
    print(f"   - Add data normalization in the CLSTM network")

if __name__ == "__main__":
    analyze_training_metrics()