requirements.txt
# Alpaca SDK
alpaca-py>=0.8.0
alpaca-trade-api>=3.0.0

# Data processing
numpy>=1.23.0
pandas>=1.5.0
scikit-learn>=1.0.0
scipy>=1.10.0

# ML frameworks
torch>=2.0.0
torchvision>=0.15.0
stable-baselines3>=2.0.0

# Technical indicators
ta-lib>=0.4.0
ta>=0.10.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# NLP models
transformers>=4.30.0
sentencepiece>=0.1.99

# Computer Vision
ultralytics>=8.0.0
pillow>=9.0.0

# Async
asyncio>=3.4.3
aiohttp>=3.8.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0
gym>=0.26.0
pyyaml>=6.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0

# Options trading
wandb>=0.15.0